requirements.txt
apipkg==1.5
atomicwrites==1.3.0
attrs==19.1.0
coverage==4.5.4
dataclasses==0.6
Django==2.2.4
django-filter==2.2.0
django-redis-cache==2.0.0
djangorestframework==3.10.2
djongo==1.2.33
execnet==1.7.0
Faker==0.9.1
importlib-metadata==0.19
Markdown==3.1.1
mixer==6.1.3
more-itertools==7.2.0
mysqlclient==1.4.2
packaging==19.1
pluggy==0.12.0
py==1.8.0
pymongo==3.8.0
pyparsing==2.4.2
pytest==5.1.0
pytest-cov==2.7.1
pytest-dependency==0.4.0
pytest-django==3.5.1
pytest-forked==1.0.2
pytest-xdist==1.29.0
python-dateutil==2.8.0
pytz==2019.2
redis==3.3.7
rest-meets-djongo==0.0.9
six==1.12.0
sqlparse==0.2.4
text-unidecode==1.2
wcwidth==0.1.7
zipp==0.5.2